requirements.txt
python-dotenv
python-redmine
langchain
langchain-community
langchain-openai
faiss-cpu
dashscope
requests
tenacity